(#101) Refactor hooks common code
